Convert CxPlot to TS
