+++ package.json
@@ -49,8 +49,10 @@
   "devDependencies": {
     "@types/jest": "^29.0.3",
     "@types/node": "^18.7.18",
+    "@types/plotly.js": "^2.12.4",
     "@types/react": "^18.0.20",
     "@types/react-dom": "^18.0.6",
+    "@types/react-plotly.js": "^2.5.2",
     "gh-pages": "4.0.0",
     "jest-canvas-mock": "2.4.0",
     "prettier": "2.7.1",
